fix: send user-left events for trailing users when a meeting ends
BBB meeting end events imply user left events for any trailing users.
This is not a thing for output modules here (eg xapi), which expect
user event chains to be completely described (ie an user join must have
a compatible user-left).

This commit spoofs user left events for remaining users in a meeting
whenever a meeting-ended event is generated for that session. There's
some extra data present in this spoofed event when compared to the
original one, but it doesn't affect functionality at all - if anything,
the original event will be extended with that extra info later on.

This also fixes an issue with user DB mappings not being cleaned up on
meeting end.