fix: handle missing checksum in API requests
API requests with missing checksums were causing an uncaught exception
that made the application return an incorrect 200 response to callers.
The expected behavior is for it to return a 200 response with a
`checksumError` key in its XML response body.

Correctly handle missing checksum when checking whether its valid on API
requests.