fix(webhooks): guarantee checksum uniqueness
When the auth method for webhooks is checksum-based, checksums are not
changing between different events for a same callback URL. The culprit
is the use of JSON.stringify to serialize the request body when
computing the checksum. The body is a URLSearchParams object, which is
not JSON.stringifiable and always return an empty object - causing the
computed checksum to never change.

Use URLSearchParams' toString method to serialize data when computing an
event's checksum. This fixes the regression and restores 2.x-era
functionality.