fix(webhooks): meetingID not respected for scoped hooks
Scoped hooks created by specifying a meetingID in the hooks/create API
call are not honored. This is caused by a typo when storing the
meetingID information in the hook DB - which makes retriaval return
nothing later on. Additionaly, the util used to check whether a hook is
global or scoped also has a typo - which affects the hooks/list API.

Fix the storage of meetingID in the Hooks DB, as well as the properly
check whether a hook is global or scoped.