fix: internal/external-meeting-id missing from some events
Some specific events may be missing internal-external-id and/or
external-meeting-id. This is caused by an inaccurate parsing of the
internal meeting ID from BBB events that have multiple base formats
across different events.

Extend the meeting ID parsing in event processing so that it covers most
of BBB message formats. This fixes a bunch of scenarios in which int/ext
meeting IDs are absent from events.